movies/requirement.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
h11==0.16.0
idna==3.11
libtorrent==2.0.11
requests==2.33.1
sqlparse==0.5.5
urllib3==2.6.3
uvicorn==0.44.0
